Remove ALCCustomInstrumentModel
