+++ qt/widgets/instrumentview/inc/MantidQtWidgets/InstrumentView/BaseCustomInstrumentModel.h
@@ -7,8 +7,10 @@
 #pragma once
 
 #include "DllOption.h"
+#include "MantidAPI/MatrixWorkspace_fwd.h"
 
 #include <map>
+#include <optional>
 #include <string>
 
 namespace MantidQt {
@@ -28,9 +30,19 @@ public:
   virtual std::string dataFileName() = 0;
   virtual int currentRun() = 0;
   virtual bool isErrorCode(const int run) = 0;
+  virtual std::map<std::string, bool> isDataValid() = 0;
   virtual const std::string getInstrument() = 0;
   virtual const std::string getTmpName() = 0;
   virtual const std::string getWSName() = 0;
+  virtual std::string WSName() = 0;
+  virtual void transformData() = 0;
+  virtual void storeSingleTube(const std::string &name) = 0;
+  virtual void extractSingleTube() = 0;
+  virtual void averageTube() = 0;
+  virtual void loadAlg(const std::string &name) = 0;
+  virtual bool averageTubeCondition(std::map<std::string, bool> tabBools) = 0;
+  virtual bool hasTubeBeenExtracted(const std::string &name) = 0;
+  virtual bool extractTubeCondition(std::map<std::string, bool> tabBools) = 0;
 };
 
 class EXPORT_OPT_MANTIDQT_INSTRUMENTVIEW BaseCustomInstrumentModel : public virtual IBaseCustomInstrumentModel {
@@ -51,8 +63,20 @@ public:
   const std::string getInstrument() override { return m_instrumentName; };
   const std::string getTmpName() override { return m_tmpName; };
   const std::string getWSName() override { return m_wsName; };
+  void transformData() override;
+  void extractSingleTube() override;
+  void averageTube() override;
+  void storeSingleTube(const std::string &name) override;
+  std::optional<double> xConversionFactor(Mantid::API::MatrixWorkspace_const_sptr workspace) const;
+  std::string WSName() override;
+  void loadAlg(const std::string &name) override;
+  std::map<std::string, bool> isDataValid() override;
+  bool averageTubeCondition(std::map<std::string, bool> tabBools) override;
+  bool hasTubeBeenExtracted(const std::string &name) override;
+  bool extractTubeCondition(std::map<std::string, bool> tabBools) override;
 
 protected:
+  int m_numberOfTubesInAverage;
   int m_currentRun;
   std::string m_tmpName;
   std::string m_instrumentName;
